+++ sdk/redisenterprise/Azure.ResourceManager.RedisEnterprise/CHANGELOG.md
@@ -1,16 +1,25 @@
 # Release History
 
-## 1.4.0-beta.1 (Unreleased)
+## 1.4.0-beta.2 (Unreleased)
 
 ### Features Added
-
-- Upgraded API version to `2025-08-01-preview`.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.4.0-beta.1 (2026-06-28)
+
+### Features Added
+
+- Upgraded API version to `2025-08-01-preview`.
+
+### Other Changes
+
+- Upgraded dependent Azure.Core to 1.59.0.
+- Upgraded dependent Azure.ResourceManager to 1.14.0.
 
 ## 1.3.0 (2025-11-10)